refactor news.js sc-5682
